+++ packages/core/_src/Effect.ts
@@ -7,6 +7,7 @@ import * as Exit from "@effect/io/Exit"
 import * as Fiber from "@effect/io/Fiber"
 import * as Layer from "@effect/io/Layer"
 import { Option } from "@fp-ts/core/Option"
+import type { ForceLazyArg } from "./_ext/pipe.ext.js"
 import { curry, flow, pipe } from "./Function.js"
 
 export * from "@effect/io/Effect"
@@ -125,7 +126,7 @@ export function tryCatchPromiseWithInterrupt<E, A>(
       promise()
         .then(x => pipe(x, Effect.succeed, resolve))
         .catch(x => pipe(x, onReject, Effect.fail, resolve))
-      return Either.left(Effect(canceller))
+      return Either.left(Effect.sync(canceller))
     })
   )
 }
@@ -322,3 +323,8 @@ export function modifyWithPermitWithEffect<A>(ref: Ref<A>, semaphore: Semaphore)
       )
     )
 }
+
+/**
+ * @tsplus static effect/io/Effect.Ops __call
+ */
+export const effectSync: <A>(evaluate: ForceLazyArg<A>) => Effect<never, never, A> = Effect.sync