+++ packages/core/_src/http/http-client-fetch.ts
@@ -28,7 +28,7 @@ function getBody(
   )
 }
 
-const makeAbort = Effect(() => new AbortController())
+const makeAbort = Effect(new AbortController())
 
 export const Client = (fetchApi: typeof fetch) =>
   H.Http.makeLayer({