+++ packages/infra-adapters/_src/express.ts
@@ -77,11 +77,11 @@ export const makeExpressApp = Effect.gen(function*(_) {
   const open = yield* _(
     Ref.make(true)
       .acquireRelease(
-        a => Effect(() => a.set(false))
+        a => Effect(a.set(false))
       )
   )
 
-  const app = yield* _(Effect(() => express()))
+  const app = yield* _(Effect(express()))
 
   const { exitHandler, host, port } = yield* _(ExpressAppConfig.access)
 
@@ -95,7 +95,7 @@ export const makeExpressApp = Effect.gen(function*(_) {
       }
       const server = app.listen(port, host, () => {
         cb(
-          Effect(() => {
+          Effect.sync(() => {
             server.removeListener("error", onError)
             return server
           })
@@ -302,7 +302,7 @@ export function match(method: Methods): {
   return function(path, ...handlers) {
     return expressRuntime(handlers).flatMap(expressHandlers =>
       withExpressApp(app =>
-        Effect(() => {
+        Effect.sync(() => {
           app[method](path, ...expressHandlers)
         })
       )
@@ -316,7 +316,7 @@ export function defaultExitHandler(
   _next: NextFunction
 ): (cause: Cause<never>) => Effect<never, never, void> {
   return cause =>
-    Effect(() => {
+    Effect.sync(() => {
       if (cause.isDie()) {
         console.error(cause.pretty)
       }
@@ -366,13 +366,13 @@ export function use(...args: any[]) {
         args as unknown as NonEmptyArguments<
           EffectRequestHandler<any, any, any, any, any, any>
         >
-      ).flatMap(expressHandlers => Effect(() => app.use(...expressHandlers)))
+      ).flatMap(expressHandlers => Effect(app.use(...expressHandlers)))
     } else {
       return expressRuntime(
         args.slice(1) as unknown as NonEmptyArguments<
           EffectRequestHandler<any, any, any, any, any, any>
         >
-      ).flatMap(expressHandlers => Effect(() => app.use(args[0], ...expressHandlers)))
+      ).flatMap(expressHandlers => Effect(app.use(args[0], ...expressHandlers)))
     }
   })
 }
@@ -414,7 +414,7 @@ export function classic(_: RequestHandler): EffectRequestHandler<never>
 export function classic(
   _: RequestHandler | NextHandleFunction
 ): EffectRequestHandler<never> {
-  return (req, res, next) => Effect(() => _(req, res, next))
+  return (req, res, next) => Effect(_(req, res, next))
 }
 
 /**