+++ packages/infra/_src/api/defaultErrorHandler.ts
@@ -20,7 +20,7 @@ export function defaultBasicErrorHandler<R>(
     r2
       .tapErrorCause(cause => cause.isFailure() ? logRequestError(cause) : Effect.unit)
       .catchTag("ValidationError", err =>
-        Effect(() => {
+        Effect.sync(() => {
           res.status(400).send(err.errors)
         }))
       // final catch all; expecting never so that unhandled known errors will show up
@@ -50,27 +50,27 @@ export function defaultErrorHandler<R>(
     r3
       .tapErrorCause(cause => cause.isFailure() ? logRequestError(cause) : Effect.unit)
       .catchTag("ValidationError", err =>
-        Effect(() => {
+        Effect.sync(() => {
           res.status(400).send(err.errors)
         }))
       .catchTag("NotFoundError", err =>
-        Effect(() => {
+        Effect.sync(() => {
           res.status(404).send(err)
         }))
       .catchTag("NotLoggedInError", err =>
-        Effect(() => {
+        Effect.sync(() => {
           res.status(401).send(err)
         }))
       .catchTag("UnauthorizedError", err =>
-        Effect(() => {
+        Effect.sync(() => {
           res.status(403).send(err)
         }))
       .catchTag("InvalidStateError", err =>
-        Effect(() => {
+        Effect.sync(() => {
           res.status(422).send(err)
         }))
       .catchTag("OptimisticConcurrencyException", err =>
-        Effect(() => {
+        Effect.sync(() => {
           // 412 or 409.. https://stackoverflow.com/questions/19122088/which-http-status-code-to-use-to-reject-a-put-due-to-optimistic-locking-failure
           res.status(412).send(err)
         }))