+++ packages/prelude/_src/utils.ts
@@ -70,7 +70,7 @@ export function Effect_debug<R, E, A>(self: Effect<R, E, A>, name: string) {
  * @tsplus fluent effect/io/Effect debugUnsafe
  */
 export function Effect_debugUnsafe<R, E, A>(self: Effect<R, E, A>, name: string) {
-  return self.tap(a => Effect(() => console.log(name, a)))
+  return self.tap(a => Effect(console.log(name, a)))
 }
 
 /**